moved: cobra cmd to pkg
