+++ pkg/cmd/cmd_destroy.go
