--- pkg/kubitect/doc.go
@@ -1,5 +1,0 @@
-// Package kubitect provides the functions used by the kubitect command line tool
-// under github.com/MusicDin/kubitect/cmd/kubitect. The functions are kept in this
-// package to better organize the code and to make it easier to test. The main() method
-// for kubitect is an only an entry point and should not contain any additional logic.
-package kubitect
